lidar-2dgs/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lidar-2dgs"
version = "1.0.0"
description = "Professional LiDAR to 2D Gaussian Splatting Converter"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "LiDAR 2DGS Team"}
]
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.3.0",
    "tqdm>=4.65.0",
    "psutil>=5.9.0",
    "laspy>=2.5.0",
    "trimesh>=3.21.0",
    "joblib>=1.3.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
gpu = [
    "torch>=2.0.0",
    "faiss-cpu>=1.7.0",
]
viewer = [
    "glfw>=2.0.0",
    "PyOpenGL>=3.1.0",
    "PyOpenGL-accelerate>=3.1.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
lidar2dgs = "tools.txt_to_2dgs:main"
lidar2dgs-large = "tools.txt_to_2dgs_large:main"
streaming-viewer = "tools.streaming_viewer_main:main"

[project.urls]
Homepage = "https://github.com/lidar-2dgs"
Repository = "https://github.com/lidar-2dgs/lidar-2dgs"
Issues = "https://github.com/lidar-2dgs/lidar-2dgs/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
exclude = ["__pycache__", "*.pyc", "tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
